fix(oauth): fix send create user requests for loop

+++ src/backend/oauth_github_service/oauth_github_app/views/oauth_github_access_resource.py
@@ -79,7 +79,7 @@ class oauthGithubAccessResourceView(View):
             self.id = str(user.id)
             self.payload['user_id'] = self.id 
             response = self.send_create_user_request_to_endpoints()
-            if response.status_code == 400:
+            if response is not None and response.status_code == 400:
                 response_data = json.loads(response.content)
                 if response_data['message'] == "Email address already registered! Try logging in.":
                     user.delete()
@@ -105,8 +105,9 @@ class oauthGithubAccessResourceView(View):
                 'http://tournament:8000/api/tournament/add_user/',]  
         for url in urls: 
             response = send_post_request(url=url, payload=self.payload, csrf_token=self.csrf_token)
-            return response
- 
+            if response.status_code == 400:
+                return response
+                
     def init_payload(self):
         self.payload = {
             'username': self.username,
